Fix limiting height for tables
The regex for styles was not adapted correctly from the attribute case.

Rewrote the limitHeight function to add the word 'disabled' to the
height attribute (instead of removing it) because I don't want to worry
about the closing quotes.

Release-Note: none
(there will be a note in a later commit)

Ticket: #7772